% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwm_weights.R
\name{calculate_bwm_weights}
\alias{calculate_bwm_weights}
\title{Calculate BWM weights (placeholder)}
\usage{
calculate_bwm_weights(bwm_criteria, bwm_best, bwm_worst)
}
\description{
Calculate BWM weights (placeholder)
}
\keyword{internal}
