% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuuzy_vikor.R, R/fuzzy_vikor.R
\name{fuzzy_vikor}
\alias{fuzzy_vikor}
\title{Fuzzy VIKOR Method}
\usage{
fuzzy_vikor(
  decision_mat,
  criteria_types,
  v = 0.5,
  weights,
  bwm_criteria,
  bwm_best,
  bwm_worst
)

fuzzy_vikor(
  decision_mat,
  criteria_types,
  v = 0.5,
  weights,
  bwm_criteria,
  bwm_best,
  bwm_worst
)
}
\arguments{
\item{decision_mat}{Matrix (m x 3n). Alternatives x fuzzy criteria.}

\item{criteria_types}{Character vector length n ("max" or "min").}

\item{v}{Numeric in [0,1]. Strategy weight.}

\item{weights}{Optional fuzzy weights (length 3n).}

\item{bwm_criteria}{Optional BWM criteria.}

\item{bwm_best}{Optional BWM best-to-others vector.}

\item{bwm_worst}{Optional BWM others-to-worst vector.}
}
\value{
Object of class `fuzzy_vikor_res`.

Object of class `fuzzy_vikor_res`.
}
\description{
Implements the Fuzzy VIKOR MCDM technique.

Implements the Fuzzy VIKOR MCDM technique.
}
