% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_vikor.R
\name{fuzzy_vikor}
\alias{fuzzy_vikor}
\title{Fuzzy VIKOR Method}
\usage{
fuzzy_vikor(
  decision_matrix,
  criteria_direction,
  strategy_weight = 0.5,
  weights,
  bwm_criteria,
  bwm_best,
  bwm_worst
)
}
\arguments{
\item{decision_matrix}{Matrix (m x 3n). Alternatives x fuzzy criteria (triangular TFN).}

\item{criteria_direction}{Character vector length n: "max" (benefit) or "min" (cost).}

\item{strategy_weight}{Numeric in [0,1]. Weight for the strategy of maximum group utility.}

\item{weights}{Optional fuzzy weights (length 3n).}

\item{bwm_criteria}{Optional BWM criteria names (if weights missing).}

\item{bwm_best}{Optional BWM best-to-others vector.}

\item{bwm_worst}{Optional BWM others-to-worst vector.}
}
\value{
Object of class `fuzzy_vikor_res`.
}
\description{
Implements the Fuzzy VIKOR MCDM technique.
}
