% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fuzzy_vikor.R
\name{plot.fuzzy_vikor_res}
\alias{plot.fuzzy_vikor_res}
\title{Bubble plot for fuzzy VIKOR results}
\usage{
\method{plot}{fuzzy_vikor_res}(x, ...)
}
\arguments{
\item{x}{object of class fuzzy_vikor_res}

\item{...}{passed to plot}
}
\description{
Bubble plot for fuzzy VIKOR results
}
